+++ kartothek/utils/__init__.py
@@ -1,0 +1,5 @@
+"""
+Utilities that are not klee-specific but are required to archive certain tasks.
+"""
+
+__all__ = ()