+++ tests/serialization/test_util.py
@@ -9,5 +9,5 @@ from kartothek.serialization._util import ensure_unicode_string_type
 )
 def test_ensure_unicode_string_types(obj, expected):
     actual = ensure_unicode_string_type(obj)
-    assert type(actual) == str
+    assert isinstance(actual, str)
     assert actual == expected